+++ cpp11test/R/cpp11.R
@@ -16,6 +16,10 @@ my_warning <- function(mystring, myarg) {
   invisible(.Call(`_cpp11test_my_warning`, mystring, myarg))
 }
 
+my_message <- function(mystring, myarg) {
+  invisible(.Call(`_cpp11test_my_message`, mystring, myarg))
+}
+
 my_stop_n1fmt <- function(mystring) {
   invisible(.Call(`_cpp11test_my_stop_n1fmt`, mystring))
 }
@@ -30,6 +34,10 @@ my_warning_n1fmt <- function(mystring) {
 
 my_warning_n2fmt <- function(mystring, myarg) {
   invisible(.Call(`_cpp11test_my_warning_n2fmt`, mystring, myarg))
+}
+
+my_message_nfmt <- function(mystring, myarg) {
+  invisible(.Call(`_cpp11test_my_message_nfmt`, mystring, myarg))
 }
 
 remove_altrep <- function(x) {