+++ cpp11test/src/errors.cpp
@@ -1,3 +1,4 @@
+#include "cpp11/function.hpp"
 #include "cpp11/protect.hpp"
 using namespace cpp11;
 
@@ -10,4 +11,7 @@ using namespace cpp11;
 }
 [[cpp11::register]] void my_warning_n2fmt(std::string mystring, std::string myarg) {
   cpp11::warning(mystring, myarg.c_str());
+}
+[[cpp11::register]] void my_message_nfmt(std::string mystring, std::string myarg) {
+  cpp11::message(mystring, myarg.c_str());
 }