+++ cpp11test/src/errors_fmt.cpp
@@ -1,4 +1,5 @@
 #define CPP11_USE_FMT
+#include "cpp11/function.hpp"
 #include "cpp11/protect.hpp"
 using namespace cpp11;
 
@@ -7,4 +8,7 @@ using namespace cpp11;
 }
 [[cpp11::register]] void my_warning(std::string mystring, std::string myarg) {
   cpp11::warning(mystring, myarg);
+}
+[[cpp11::register]] void my_message(std::string mystring, std::string myarg) {
+  cpp11::message(mystring, myarg);
 }