+++ cpp11test/tests/testthat/test_formatted_errors.R
@@ -13,6 +13,13 @@ test_that("cpp11::warning formatting works", {
   test2 <- c("failed", "passed")
   expect_warning(my_warning("You {}", test2[2]), "You passed", fixed = TRUE)
 })
+test_that("cpp11::message formatting works", {
+  test1 <- "message"
+  expect_message(my_message("This is a {}", test1), "This is a message", fixed = TRUE)
+
+  test2 <- c("great", "super")
+  expect_message(my_message("You're {}", test2[2]), "You're super", fixed = TRUE)
+})
 test_that("cpp11::stop works without including the fmt library", {
   test1 <- "error"
   expect_error(my_stop_n1fmt("This is a stop"), "This is a stop", fixed = TRUE)
@@ -22,4 +29,11 @@ test_that("cpp11::warning works without including the fmt library", {
   test1 <- "warning"
   expect_warning(my_warning_n1fmt("This is a warning"), "This is a warning", fixed = TRUE)
   expect_warning(my_warning_n2fmt("This is a %s", test1), "This is a warning", fixed = TRUE)
+})
+test_that("cpp11::message works without including the fmt library", {
+  test1 <- "message"
+  expect_message(my_message_nfmt("This is a %s", test1), "This is a message", fixed = TRUE)
+
+  test2 <- c("great", "super")
+  expect_message(my_message_nfmt("You're %s", test2[2]), "You're super", fixed = TRUE)
 })