+++ NEWS.md
@@ -5,6 +5,7 @@
 * Error messages now output original file name rather than the temporary file name (@sbearrows, #194)
 * Fixed bug when running `cpp_source()` on the same file more than once (@sbearrows, #202)
 * Removed internal instances of `cpp11::stop()` and replaced with C++ exceptions (@sbearrows, #203)
+* Added `cpp11::messages()` feature with {fmt} library formatting for non-error messages (@sbearrows, #208)
 * Names of named lists are now resized along with the list elements (@sbearrows, #206)
 * Added optionally formatting to `stop()` and `warning()` using {fmt} library (@sbearrows, #169)